9043/refactor: extract process WayForPaySystem and result handler for WayForPay
